chore: disable auto-merge on human push to dependabot PR
When dependabot PR fails and a human starts interacting with it to try
to unblock it, it would previously be left with auto-merge enabled,
which would surprise the human if, for example, they pushed an
experimental change meant to unblock CI, and it accidentally ended up on
main. Therefore, if an author besides dependabot pushes to the
dependabot PR, disable the auto-merge.

Signed-off-by: Will Murphy <will.murphy@anchore.com>